Refactor worker into platform-based components
